+++ internal/compiler/generator/cpp.rs
@@ -2794,6 +2794,18 @@ fn compile_builtin_function_call(
                 panic!("internal error: invalid args to ItemMemberFunction {:?}", arguments)
             }
         }
+        BuiltinFunction::MapPointToWindow => {
+            if let [llr::Expression::PropertyReference(pr), llr::Expression::Struct { values, .. }] =
+                arguments
+            {
+                let item_rc = access_item_rc(pr, ctx);
+                let x_expr = compile_expression(values.get("x").unwrap(), ctx);
+                let y_expr = compile_expression(values.get("y").unwrap(), ctx);
+                format!("slint_item_map_point_to_window(&{item_rc}, {x_expr}, {y_expr})")
+            } else {
+                panic!("internal error: invalid args to MapPointToWindow {:?}", arguments)
+            }
+        }
         BuiltinFunction::RegisterCustomFontByPath => {
             if let [llr::Expression::StringLiteral(path)] = arguments {
                 let window = access_window_field(ctx);